openai_api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
distro==1.8.0
exceptiongroup==1.1.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
openai==1.2.4
packaging==23.2
pluggy==1.3.0
pydantic==2.5.0
pydantic_core==2.14.1
pytest==7.4.3
python-dotenv==1.0.0
sniffio==1.3.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
